+++ tabbycat/settings/__init__.py
@@ -1,7 +1,17 @@
 import os
+import logging
+import sys
 
 from split_settings.tools import optional, include
 
+
+root = logging.getLogger()
+root.setLevel(logging.DEBUG)
+ch = logging.StreamHandler(sys.stdout)
+ch.setLevel(logging.DEBUG)
+formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+ch.setFormatter(formatter)
+root.addHandler(ch)
 
 # ==============================================================================
 # Environments
@@ -9,20 +19,26 @@ from split_settings.tools import optional, include
 
 base_settings = [
     'core.py',
-    optional('local.py'),
 ]
 
-if os.environ.get('CI', '') == 'true':
+if os.environ.get('GITHUB_CI', '') and bool(os.environ['GITHUB_CI']):
     base_settings.append('github.py')
-
-if os.environ.get('IN_DOCKER', '') and bool(int(os.environ['IN_DOCKER'])):
+    root.info('SPLIT_SETTINGS: imported github.py')
+elif os.environ.get('IN_DOCKER', '') and bool(int(os.environ['IN_DOCKER'])):
     base_settings.append('docker.py')
-
-if os.environ.get('ON_HEROKU', ''):
+    root.info('SPLIT_SETTINGS: imported github.py')
+elif os.environ.get('ON_HEROKU', ''):
     base_settings.append('heroku.py')
-
-if os.environ.get('LOCAL_DEVELOPMENT', ''):
-    base_settings.append('development.py')
-
+    root.info('SPLIT_SETTINGS: imported heroku.py')
+elif os.environ.get('ON_RENDER', ''):
+    base_settings.append('render.py')
+    root.info('SPLIT_SETTINGS: imported render.py')
+else:
+    base_settings.append('local.py')
+    if os.environ.get('LOCAL_DEVELOPMENT', ''):
+        base_settings.append('development.py')
+        root.info('SPLIT_SETTINGS: imported local.py & development.py')
+    else:
+        root.info('SPLIT_SETTINGS: imported local.py')
 
 include(*base_settings)